Add Declarative Shadow DOM to list of good explainers. (w3ctag/tag.w3.org#27)
@alice suggested in today's breakout B that the Declarative Shadow DOM
explainer is a good explainer and should be added to the list of good
explainers.  I agree with this suggestion, so here's a PR adding it.

(Note that saying the explainer being good isn't necessarily saying that
I like the feature.  I think I still disagree with the tradeoffs made
here... however, I agree that the explainer does a good job of
explaining the feature and it makes sense to link to it.)